+++ src/main/scala/sparkz/core/network/peer/ForgerNodePeerFeature.scala
@@ -1,0 +1,34 @@
+package sparkz.core.network.peer
+
+import sparkz.core.network.PeerFeature
+import sparkz.core.network.PeerFeature.Id
+import sparkz.util.serialization._
+import sparkz.core.serialization.SparkzSerializer
+
+/**
+  * This peer feature marks forger nodes
+  */
+case class ForgerNodePeerFeature() extends PeerFeature {
+
+  override type M = ForgerNodePeerFeature
+  override val featureId: Id = ForgerNodePeerFeature.featureId
+
+  override def serializer: ForgerNodePeerFeatureSerializer.type = ForgerNodePeerFeatureSerializer
+
+}
+
+object ForgerNodePeerFeature {
+
+  val featureId: Id = 5: Byte
+
+}
+
+object ForgerNodePeerFeatureSerializer extends SparkzSerializer[ForgerNodePeerFeature] {
+
+
+  override def parse(r: Reader): ForgerNodePeerFeature = {
+    ForgerNodePeerFeature()
+  }
+
+  override def serialize(obj: ForgerNodePeerFeature, w: Writer): Unit = {}
+}