+++ src/test/scala/sparkz/core/network/NetworkControllerSpec.scala
@@ -18,7 +18,7 @@ import sparkz.core.network.NetworkController.ReceivableMessages.{ConnectTo, GetC
 import sparkz.core.network.NodeViewSynchronizer.ReceivableMessages.DisconnectedPeer
 import sparkz.core.network.message._
 import sparkz.core.network.peer.PeerManager.ReceivableMessages.{AddOrUpdatePeer, ConfirmConnection, DisconnectFromAddress, GetAllPeers, RandomPeerForConnectionExcluding}
-import sparkz.core.network.peer._
+import sparkz.core.network.peer.{ForgerNodePeerFeature, _}
 import sparkz.core.serialization.SparkzSerializer
 import sparkz.core.settings.SparkzSettings
 import sparkz.core.utils.LocalTimeProvider
@@ -33,7 +33,8 @@ class NetworkControllerSpec extends NetworkTests with ScalaFutures {
   import scala.concurrent.ExecutionContext.Implicits.global
 
   private val featureSerializers = Map[Byte, SparkzSerializer[_ <: PeerFeature]](LocalAddressPeerFeature.featureId -> LocalAddressPeerFeatureSerializer,
-    TransactionsDisabledPeerFeature.featureId -> TransactionsDisabledPeerFeatureSerializer)
+    TransactionsDisabledPeerFeature.featureId -> TransactionsDisabledPeerFeatureSerializer,
+    ForgerNodePeerFeature.featureId -> ForgerNodePeerFeatureSerializer)
 
   "A NetworkController" should "send local address on handshake when peer and node address are in localhost" in {
     implicit val system: ActorSystem = ActorSystem()
@@ -264,7 +265,7 @@ class NetworkControllerSpec extends NetworkTests with ScalaFutures {
     val tcpManagerProbe = TestProbe()
 
     val nodeAddr = new InetSocketAddress("88.77.66.55", 12345)
-    val settings2 = settings.copy(network = settings.network.copy(bindAddress = nodeAddr, maxOutgoingConnections = 1))
+    val settings2 = settings.copy(network = settings.network.copy(bindAddress = nodeAddr, maxOutgoingConnections = 1, maxForgerConnections = 0))
     val (networkControllerRef: ActorRef, _) = createNetworkController(settings2, tcpManagerProbe)
 
     val testPeer = new TestPeer(settings2, networkControllerRef, tcpManagerProbe)
@@ -282,6 +283,70 @@ class NetworkControllerSpec extends NetworkTests with ScalaFutures {
     val peerInfo2 = getPeerInfo(peer2LocalAddr)
     testPeer.establishNewOutgoingConnection(peerInfo2)
     testPeer.connectAndExpectMessage(peer2LocalAddr, nodeAddr, Tcp.Close)
+
+    system.terminate()
+  }
+
+  it should "allow connecting to forger over maxOutgoingConnections limit" in {
+    implicit val system: ActorSystem = ActorSystem()
+
+    val tcpManagerProbe = TestProbe()
+
+    val nodeAddr = new InetSocketAddress("88.77.66.55", 12345)
+    val settings2 = settings.copy(network = settings.network.copy(bindAddress = nodeAddr, maxOutgoingConnections = 1, maxForgerConnections = 1))
+    val (networkControllerRef: ActorRef, _) = createNetworkController(settings2, tcpManagerProbe)
+
+    val testPeer = new TestPeer(settings2, networkControllerRef, tcpManagerProbe)
+    val peer1DecalredAddr = new InetSocketAddress("88.77.66.55", 5678)
+    val peer1LocalAddr = new InetSocketAddress("192.168.1.55", 5678)
+    val peerInfo1 = getPeerInfo(peer1LocalAddr)
+    testPeer.establishNewOutgoingConnection(peerInfo1)
+    testPeer.connectAndExpectSuccessfulMessages(peer1LocalAddr, nodeAddr, Tcp.ResumeReading)
+    testPeer.receiveHandshake
+    testPeer.sendHandshake(Some(peer1DecalredAddr), Some(peer1LocalAddr))
+    testPeer.receiveGetPeers
+    testPeer.sendPeers(Seq.empty)
+
+    val peer2DeclaredAddr = new InetSocketAddress("99.88.77.66", 5678)
+    val peer2LocalAddr = new InetSocketAddress("192.168.1.56", 5678)
+    val peerInfo2 = getPeerInfo(peer2LocalAddr)
+    testPeer.establishNewOutgoingConnection(peerInfo2)
+    testPeer.connectAndExpectSuccessfulMessages(peer2LocalAddr, nodeAddr, Tcp.ResumeReading)
+    testPeer.receiveHandshake
+    testPeer.sendHandshake(Some(peer2DeclaredAddr), Some(peer2LocalAddr), forgerNode = true)
+    tcpManagerProbe.expectNoMessage()
+
+    system.terminate()
+  }
+
+  it should "disconnect after handshake if the maxOutgoingConnections limit reached" in {
+    implicit val system: ActorSystem = ActorSystem()
+
+    val tcpManagerProbe = TestProbe()
+
+    val nodeAddr = new InetSocketAddress("88.77.66.55", 12345)
+    val settings2 = settings.copy(network = settings.network.copy(bindAddress = nodeAddr, maxOutgoingConnections = 1, maxForgerConnections = 1))
+    val (networkControllerRef: ActorRef, _) = createNetworkController(settings2, tcpManagerProbe)
+
+    val testPeer = new TestPeer(settings2, networkControllerRef, tcpManagerProbe)
+    val peer1DecalredAddr = new InetSocketAddress("88.77.66.55", 5678)
+    val peer1LocalAddr = new InetSocketAddress("192.168.1.55", 5678)
+    val peerInfo1 = getPeerInfo(peer1LocalAddr)
+    testPeer.establishNewOutgoingConnection(peerInfo1)
+    testPeer.connectAndExpectSuccessfulMessages(peer1LocalAddr, nodeAddr, Tcp.ResumeReading)
+    testPeer.receiveHandshake
+    testPeer.sendHandshake(Some(peer1DecalredAddr), Some(peer1LocalAddr))
+    testPeer.receiveGetPeers
+    testPeer.sendPeers(Seq.empty)
+
+    val peer2DeclaredAddr = new InetSocketAddress("99.88.77.66", 5678)
+    val peer2LocalAddr = new InetSocketAddress("192.168.1.56", 5678)
+    val peerInfo2 = getPeerInfo(peer2LocalAddr)
+    testPeer.establishNewOutgoingConnection(peerInfo2)
+    testPeer.connectAndExpectSuccessfulMessages(peer2LocalAddr, nodeAddr, Tcp.ResumeReading)
+    testPeer.receiveHandshake
+    testPeer.sendHandshake(Some(peer2DeclaredAddr), Some(peer2LocalAddr))
+    tcpManagerProbe.expectMsg(Abort)
 
     system.terminate()
   }
@@ -626,7 +691,8 @@ class TestPeer(settings: SparkzSettings, networkControllerRef: ActorRef, tcpMana
 
   private val timeProvider = LocalTimeProvider
   private val featureSerializers = Map[Byte, SparkzSerializer[_ <: PeerFeature]](LocalAddressPeerFeature.featureId -> LocalAddressPeerFeatureSerializer,
-    TransactionsDisabledPeerFeature.featureId -> TransactionsDisabledPeerFeatureSerializer)
+    TransactionsDisabledPeerFeature.featureId -> TransactionsDisabledPeerFeatureSerializer,
+    ForgerNodePeerFeature.featureId -> ForgerNodePeerFeatureSerializer)
   private val handshakeSerializer = new HandshakeSpec(featureSerializers, Int.MaxValue)
   private val peersSpec = new PeersSpec(featureSerializers, settings.network.maxPeerSpecObjects)
   private val messageSpecs = Seq(GetPeersSpec, peersSpec)
@@ -675,9 +741,10 @@ class TestPeer(settings: SparkzSettings, networkControllerRef: ActorRef, tcpMana
     * @param localAddress
     * @return
     */
-  def sendHandshake(declaredAddress: Option[InetSocketAddress], localAddress: Option[InetSocketAddress]): Tcp.ResumeReading.type = {
+  def sendHandshake(declaredAddress: Option[InetSocketAddress], localAddress: Option[InetSocketAddress], forgerNode: Boolean = false): Tcp.ResumeReading.type = {
     val localFeature: Seq[PeerFeature] = localAddress.map(LocalAddressPeerFeature(_)).toSeq
-    val features = localFeature :+ SessionIdPeerFeature(settings.network.magicBytes)
+    val forgerNodeFeature = if (forgerNode) Some(ForgerNodePeerFeature()) else None
+    val features = localFeature ++ forgerNodeFeature :+ SessionIdPeerFeature(settings.network.magicBytes)
     val handshakeToNode = Handshake(PeerSpec(settings.network.agentName,
       Version(settings.network.appVersion), "test",
       declaredAddress, features), timeProvider.time())