+++ src/test/scala/sparkz/core/network/SyncTrackerSpec.scala
@@ -5,7 +5,7 @@ import akka.testkit.TestProbe
 import org.mockito.MockitoSugar.mock
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.propspec.AnyPropSpec
-import sparkz.core.consensus.History.{Fork, HistoryComparisonResult}
+import sparkz.core.consensus.History.{Fork, HistoryComparisonResult, Unknown}
 import sparkz.core.settings.NetworkSettings
 import sparkz.core.utils.NetworkTimeProvider
 import sparkz.core.utils.TimeProvider.Time
@@ -60,5 +60,22 @@ class SyncTrackerSpec extends AnyPropSpec with Matchers {
     // Assert
     statusAfterDeletion.size should be(0)
     timestampAfterDeletion.size should be(0)
+  }
+
+  property("syncTracker should start tracking a peer once the handshake message has been received") {
+    // Arrange
+    val syncTracker = new SyncTracker(probe.ref, context, networkSettings, timeProvider)
+
+    val lastSyncSentTimeField = classOf[SyncTracker].getDeclaredField("lastSyncSentTime")
+    lastSyncSentTimeField.setAccessible(true)
+
+    val lastSyncSentTime = lastSyncSentTimeField.get(syncTracker).asInstanceOf[mutable.Map[ConnectedPeer, Time]]
+
+    // Act
+    lastSyncSentTime should be(Map.empty)
+    syncTracker.updateStatus(connectedPeer, Unknown)
+
+    // Assert
+    lastSyncSentTime.contains(connectedPeer) should be(true)
   }
 }