+++ release-notes.md
@@ -1,3 +1,12 @@
+2.1.0
+---------
+* Added support for forger's connections prioritization
+* Node synchronization improvements:
+  * Optimization of lookup strategy in modifiersCache
+  * Preserve the order of block during synchronization
+* Added option to force only connecting to known peers
+* Fixes/Improvements on the way the SyncTracker handles the internal statuses maps
+
 2.0.3
 ---------
 * Fix in the handshake process - start connecting to nodes only after the Synchronizer is initialized